requirements.txt
langchain 
groq
langchain-groq
streamlit 
python-dotenv